♻ Organize discovery-based classes
